Move XCUIElementQuery+Sequence logic to UITestHelpers
